+++ test/test_asgi.py
@@ -16,8 +16,9 @@ def schema(fastapi_app):
     return from_asgi("/openapi.json", fastapi_app, force_schema_version="30")
 
 
+@pytest.mark.parametrize("method", ("call", "call_asgi"))
 @pytest.mark.hypothesis_nested
-def test_cookies(fastapi_app):
+def test_cookies(fastapi_app, method):
     @fastapi_app.get("/cookies")
     def cookies(token: str = Cookie(None)):
         return {"token": token}
@@ -50,7 +51,7 @@ def test_cookies(fastapi_app):
     @given(case=strategy)
     @settings(max_examples=3, suppress_health_check=[HealthCheck.filter_too_much], deadline=None)
     def test(case):
-        response = case.call_asgi()
+        response = getattr(case, method)()
         assert response.status_code == 200
         assert response.json() == {"token": "test"}
 
@@ -81,7 +82,7 @@ def test_null_byte(fastapi_app):
         max_examples=50, suppress_health_check=[HealthCheck.filter_too_much], deadline=None, phases=[Phase.generate]
     )
     def test(case):
-        response = case.call_asgi()
+        response = case.call()
         assert response.status_code == 200
         assert response.json() == {"success": True}
 
@@ -116,13 +117,13 @@ def test_base_url():
     def read_root():
         return {"Hello": "World"}
 
-    schema = schemathesis.from_dict(raw_schema)
+    schema = schemathesis.from_dict(raw_schema, app=app)
     strategy = schema["/foo"]["GET"].as_strategy()
 
     @given(case=strategy)
     @settings(max_examples=1, suppress_health_check=[HealthCheck.filter_too_much], deadline=None)
     def test(case):
-        response = case.call_asgi(app)
+        response = case.call()
         # Then the base path should be respected and calls should not lead to 404
         assert response.status_code == 200
 
@@ -212,7 +213,7 @@ def test_events(setup):
     @given(case=schema["/health"]["GET"].as_strategy())
     @settings(max_examples=3, deadline=None)
     def test(case):
-        response = case.call_asgi()
+        response = case.call()
         assert response.status_code == 200
         assert response.json() == {"status": "OK"}